requirements.txt
streamlit
pandas
openpyxl
pyyaml
pydantic
